Rename bind to bindContent in listener.js
This is partly to avoid confusion with the skooma `bind` function,
but also to make it clearer that only the inner content of an element is
being bound to the listener.

This function may also disappear entirely in the future if it turns out
it doesn't provide any benefits over the skooma bind function.